backend/requirements.txt
python-dotenv
flask
requests
gunicorn
google-auth
six==1.16.0
cohere